Echo more than just one message before exiting
If the process started by `echo.escript` receives a message from
anything prior to that of the test that launched it, then it will
exit, and the test will fail because its message will never be echoed.

This change prevents that scenario by moving the `receive` statement
into a loop. Also explicitly terminate `echo.escript` by sending a
`stop` message.