Avoid debug statement perf hit
The `log.debug("pids %s", processes.keys.toList)` line was showing up in a few
stack traces when clouseau was timing out connecting to the erlang VM. The
debug log line always translates the process map to a list, regardless if we're
debugging or not.

To fix it, check for debug flag and run `toList()` only when debugging.